requirements.txt
# ============================================
# SMART REPORTS - DEPENDENCIES
# Instituto Hutchison Ports v2.0
# ============================================

# Base de datos
pyodbc>=5.0.0
mysql-connector-python>=8.0.33

# Procesamiento de datos
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0

# Validacion de datos
pydantic>=2.0.0

# Interfaz grafica - PyQt6 (MIGRACIÓN COMPLETA desde CustomTkinter)
# PyQt6 ofrece mejor rendimiento, control profesional y QWebEngineView para D3.js
PyQt6>=6.6.0
PyQt6-WebEngine>=6.6.0
pillow>=10.0.0

# DEPRECADO - CustomTkinter y dependencias Tkinter ya no se usan
# Se eliminaron en favor de PyQt6 para mejor integracion con JavaScript/D3.js
# customtkinter>=5.2.0  # Removido
# tkinterdnd2-universal>=1.7.3  # Removido
# tkhtmlview>=0.1.1  # Removido
# tkinterweb>=3.23.0  # Removido - Reemplazado por QWebEngineView
# pywebview>=4.0.0  # Removido - Reemplazado por QWebEngineView

# Graficos y visualizacion
matplotlib>=3.7.0
plotly>=5.18.0
kaleido>=0.2.1

# Reportes PDF
reportlab>=4.0.0

# Utilidades
python-dateutil>=2.8.2
colorama>=0.4.6
colorlog>=6.7.0
python-dotenv>=1.0.0

# Testing (opcional)
pytest>=7.4.0
pytest-cov>=4.1.0

# Empaquetado
pyinstaller>=6.0.0

# Dependencias adicionales para CEF browser (opcional)
# cefpython3>=66.0  # Comentado - puede causar problemas de instalación en algunos sistemas

# ============================================================================
# IMPORTANTE: SQL Server ODBC Driver
# ============================================================================
# Para Linux:
# sudo apt-get install unixodbc unixodbc-dev
# curl https://packages.microsoft.com/keys/microsoft.asc | sudo apt-key add -
# curl https://packages.microsoft.com/config/ubuntu/$(lsb_release -rs)/prod.list | sudo tee /etc/apt/sources.list.d/mssql-release.list
# sudo apt-get update
# sudo ACCEPT_EULA=Y apt-get install -y msodbcsql17
#
# Para Windows:
# Descargar e instalar: https://go.microsoft.com/fwlink/?linkid=2249004
#
# Para macOS:
# brew install unixodbc
# brew tap microsoft/mssql-release https://github.com/Microsoft/homebrew-mssql-release
# brew install msodbcsql17
